Move is_docstring to misc module
